Problem: Name parse unfairly hidden in function
Solution: Bring the name parsing into the primary function

This allows future improvements to the name parsing logic.
It also means we parse the possible functions in a single big list
rather than needing to descend, which amkes future augmentations to this
procedure possible instead of very difficult.

Sometimes abstraction is premature.