Problem: Only POSIX shell scripts supported
Solution: Allow anything where the interpreter is elsewhere.

We can't execute in /dev/shm (mounted noexec normally) so we should
prepare to always require an interpreter.

Maybe though there is an interpreter that can exec things that are
noexec, by hoisting the data into a different space in memory? Musings
for another day.